+++ apps/pages/urls.py
@@ -3,6 +3,7 @@ from django.urls import path
 from apps.pages import views
 
 urlpatterns = [
+    path("mcp-server", views.McpView.as_view(), name="mcp"),
     path("privacy-policy", views.PrivacyPolicyView.as_view(), name="privacy_policy"),
     path("terms-of-service", views.TermsOfServiceView.as_view(), name="terms_of_service"),
 ]